Property class __get__, __set__, and fixed memory leak (see details)
Property values should be stored on the entity, not the property as a
property’s lifetime is the same as the model class. Meaning that if
values are stored on an entity they are removed with that entity. This
is much safer and more fitting towards a property value’s purpose